+++ demo/headless-tiny.html
@@ -9,7 +9,7 @@
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
 			import '@brightspace-ui/core/components/dialog/dialog.js';
-			import '../htmleditor-wc.js';
+			import '../htmleditor-headless.js';
 			import { html, LitElement } from 'lit-element/lit-element.js';
 
 			class DemoWrapper extends LitElement {
@@ -35,7 +35,7 @@
 		</style>
 	</head>
 	<body unresolved>
-		<d2l-demo-page page-title="d2l-htmleditor (shadowDOM enabled)">
+		<d2l-demo-page page-title="d2l-htmleditor (headless tinymce)">
 
 			<h2>Default</h2>
