+++ demo/standard-tiny.html
@@ -2,14 +2,10 @@
 <html lang="en">
 	<head>
 		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
-		<script>
-			// tinyMCE currently does not work in shadowDOM context
-			window.ShadyDOM = {force: true};
-		</script>
 		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../htmleditor.js';
+			import '../htmleditor-standard.js';
 			import { html, LitElement } from 'lit-element/lit-element.js';
 
 			class DemoWrapper extends LitElement {
@@ -35,7 +31,7 @@
 		</style>
 	</head>
 	<body unresolved>
-		<d2l-demo-page page-title="d2l-htmleditor (shadowDOM disabled)">
+		<d2l-demo-page page-title="d2l-htmleditor (standard tinymce)">
 
 			<h2>Default</h2>
