+++ package.json
@@ -27,25 +27,25 @@
   "devDependencies": {
     "@brightspace-ui/visual-diff": "^1",
     "@open-wc/testing": "^2",
-    "@open-wc/testing-karma": "^3",
+    "@open-wc/testing-karma": "^4",
     "babel-eslint": "^10",
     "deepmerge": "^4.2.2",
     "es-dev-server": "^1",
     "eslint": "^7",
-    "eslint-config-brightspace": "^0.8.0",
+    "eslint-config-brightspace": "^0.9.0",
     "eslint-plugin-html": "^6",
     "eslint-plugin-lit": "^1",
     "eslint-plugin-sort-class-members": "^1",
     "frau-ci": "^1",
     "karma-sauce-launcher": "^4.1.4",
     "lit-analyzer": "^1",
-    "mocha": "^7",
-    "puppeteer": "^2"
+    "mocha": "~8.0.1",
+    "puppeteer": "^5"
   },
   "dependencies": {
     "@brightspace-ui/core": "^1",
     "@webcomponents/webcomponentsjs": "^2",
     "lit-element": "^2",
-    "tinymce": "5.2.2"
+    "tinymce": "5.4.1"
   }
 }
